app.py: EXPLAIN needs to be executed on the good server
Since {T195836: Cannot reliably get the EXPLAIN for
a query on analytics wiki replica cluster}
it's not possible to directly get the EXPLAIN for
a query on analytics wiki replica cluster
since multiple servers are handling requests and
it's not possible to choose on which one
the EXPLAIN need to be executed.

With this patch the EXPLAIN the command is
executed while it fails up to five times or until
the two hosts in the cluster are tested.

Bug: T205214
Change-Id: I6eed037170ca717e3c3546e608a1a7362a9d127e