db: Added a script to generate a DB schema from the models
It might be simpler in the future to use flask-sqlalchemy, but did not
want to touch the current dependencies.

Had to add a length to the task_id on queryrun as it's mandatory for
MySQL to have it.
The value has been taken from the current table definiton in
production.

This is being used to start syncing the code with the actual DB.

Bug: T288523
Change-Id: I955b4b83325d41de5d0dd9549d06e83b46bd5aef