tox
This is an offshoot of the blubber implimentation. The primary element that it
lacks is the ability to run tox without needing to install redis and a specific
version of python first. It may be worth creating a container that does the
whole test mounting the local directory as it goes, basically the blubber
container, but have it in a registery so that it can be pulled.

Bug: T317092