Improve error handling by capturing all exceptions in worker.py
In an effort to facilitate debugging, the codebase now captures all exceptions rather than specifically pymysql.err.OperationalError. Alongside this, an error's traceback is being formatted and logged for additional context about the error's origin. Furthermore, the additional information pertaining to the error is now being serialized and stored into the 'extra_info' field of the 'qrun' object.

Bug: T365136