Fix bug causing wiki replica queries going to ToolsDb
The Replica object is instantiated just once per worker, as opposed to once per query run that I previously thought was the case.

This means self.is_tools_db needs to be explicitly set to `False` when we are querying wiki replicas, as it may have previously been set to `True` while querying tools db.

Bug: T365374