Enable pool_pre_ping to eliminate db connection reset errors (#51)
Enable pool_pre_ping to eliminate db connection reset errors

This sends a "SELECT 1" to the Trove db every time a db connection is retrieved from the pool. Although it adds slight latency, it is common in db pooling setups and should be worth it as it would eliminate all connection reset errors.

Also set pool_size=12, increasing it from the default 5. We have 10 pods (8 web + 2 worker), so total 120 connections which is very reasonable, as Quarry is the only user of this db.

Bug: T367464