Support querying quarry_p database
quarry_p is a new database containing read-only views mirroring 4 tables
from Quarry's own database:
- query
- query_revision
- query_run
- star

The use case for this view is described in T367415: doing stats on query
execution times, and on which tables are being queried.

These views should not expose any information that is not already
visible from the public Quarry web interface. The "user" and
"user_group" table do not have corresponding views, to preserve the
privacy of users that have logged in to Quarry but did not create any
queries.

Bug: T367415