metrics: Replace custom concatenation logic with two endpoints
The current logic tries to concatenate responses from
prometheus-flask-exporter (providing metrics for that individual app
instance) and for the custom Quarry metrics (which reference things from
the database and so are global). That does not work, since unlike the
simple text format, the binary OTel format used in production does not
allow concatenating responses like that. In addition the model where
those two are combined was designed when there only was a single
instance of Quarry so scraping both worker-specific and global metrics
from the same endpoint was acceptable.

Instead split the app to use two endpoints; /metrics continues to be the
prometheus-flask-exporter one that should be scraped per worker
(although I'm not quite sure how to do that with metricsinfra and the
Trove cluster) and /app/metrics provides the app-level metrics that
should only have one Prometheus target for it.

Additionally, by scraping the app metrics via the quarry.wmcloud.org
service name and proxy, we can use the up metric for that target to
provide alerting in case Quarry is entirely down for users.

Bug: T392138
Change-Id: Id4aa84d43976630cf722c0932937e8ddd85b07b1