helm: Use an emptyDir for /tmp (#80)
* helm: Set up a separate emptyDir for /tmp in the workers

When we recently saw the worker nodes filling up their disks, one cause
was the Quarry workers filling up their temporary directories. Set up
the temporary directory to be a separate volume with a maximum size to
prevent that from rendering the entire node unusable.

Change-Id: I7297ea17cf07baa2809951ea4fbebced55675010

* helm: Use a read-only root file system

With /tmp now being a separate volume, there's no need for anything to
write to the container root file system.

Change-Id: I9c581d082fdd98dcd224a03b6352e3b5c0ae87f0

* auto update of  tag

---------

Co-authored-by: Github Action <auto@github.com>