output: Run Excel export in a specific temporary directory
Currently Quarry is leaking temporary files that seem to be caused by
this. My testing indicates that the last commit which upgraded
xlsxwriter fixed that, but just to be sure, run the export in a
temporary directory that will be cleaned up afterwards instead of doing
it in /tmp which has a possibility of accumulating leaked files.

Bug: T395237
Change-Id: I7dc3ad3a7b638634e156f9350ea8b278ad64870b