Fix xlsx export tmpfile leaks (#82)
* build: Upgrade xlsxwriter

Bug: T395237
Change-Id: Ib3a1640ffeef6db53901b07c6ab3bf642ab53fd4

* output: Run Excel export in a specific temporary directory

Currently Quarry is leaking temporary files that seem to be caused by
this. My testing indicates that the last commit which upgraded
xlsxwriter fixed that, but just to be sure, run the export in a
temporary directory that will be cleaned up afterwards instead of doing
it in /tmp which has a possibility of accumulating leaked files.

Bug: T395237
Change-Id: I7dc3ad3a7b638634e156f9350ea8b278ad64870b

* auto update of  tag

---------

Co-authored-by: Github Action <auto@github.com>